+++ logout.php
@@ -5,4 +5,5 @@ $_SESSION['user_id'] = null;
 $_SESSION['email'] = null;
 $_SESSION['user_role'] = null;
 
-header("Location: menu_list.php");
+header("Location: index.php");
+exit();
